pyproject.toml
[project]
name = "blackjack-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.7.0",
    "litellm>=1.74.7",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dealer_agent"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
